sdk/relay-python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commandless-relay"
version = "0.1.10"
description = "Python SDK for Commandless relay API"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Commandless" }
]
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
discord = ["discord.py>=2.4.0"]

[project.scripts]
commandless-discord = "commandless_relay.discord_runner:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
